requirements.txt
cryptography>=41.0.0
websockets>=12.0
PyJWT>=2.8.0
aiofiles>=23.2.1